requirements.txt
openai>=1.0.0
anthropic>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pyyaml>=6.0
gitpython>=3.1.0
jinja2>=3.1.0
typer>=0.9.0
rich>=13.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0